Fixed additional pin reconstruction in promotable operator nodes.
If a promotable operator node had an additional pin (eg: pin C), as well as split pins for any of the default pins (eg: B_X for a vector pin), then ReallocatePinsDuringReconstruction would incorrectly identify a subpin as an additional pin. In this case, the true additional pin would get the wrong pin type.

In practice, this didn't really lead to any issues. However, a future change will make pin type checking a bit stricter, so we need to ensure that recreated pins accurately match their expected types.

The fix here simply modifies GetAdditionalPin to check that the pin isn't a subpin of a default pin.

#jira none
#preflight 63d15609574ab9cae4bbdf22
#rb ben.hoffman

[CL 23851005 by dave jones2 in ue5-main branch]