Apparently, ExpandEnumAsExecs and ExpandBoolAsExecs have some overlap, and this change was a bit too restrictive for the latter case. This ended up causing a pin lookup to fail in Lyra.
[Backout] - CL23820159
#fyi dave.jones2
Original CL Desc
-----------------------------------------------------------------
UE-174107 - ExpandEnumAsExecs always uses first Exec if an argument name matches Enum value

ExpandEnumAsExecs was using an ambiguous pin lookup. These types of function nodes add additional exec pins that are named after the enumerators. If the target pin specified by ExpandEnumAsExecs happens to have the same name as an enumerator value, then we'll end up with the wrong pin if the basic FindPin function is used.

Instead, we prefer using FindPinByPredicate to specify the exact pin that we're interested in. This ensures that the expansion step creates a SwitchEnum node with a valid UEnum object.

#jira UE-174107
#preflight 63cb0789fa290562c1c8b6b2
#rb phillip.kavan

[CL 23859234 by dave jones2 in ue5-main branch]