Significantly improved dragging performance for blueprints with many static meshes, in both level and blueprint editor.  Especially large performance wins occur when dragging many items in a blueprint editor with a level editor open.  With "Run Construction Script on Drag" enabled, move cost goes from 9040 ms to 89 ms, in a test level moving 100 items in a 200 item blueprint.  With Construction Script disabled, the same move cost goes from 1106 ms to 38 ms.  For moving a 200 item blueprint in the level editor, performance is 48% faster (89 ms -> 46 ms) with Construction Script enabled.  With Construction Script disabled, performance is 36% faster for a 1200 item blueprint (42 ms -> 27 ms).
The reason for the especially large performance gains is due to a couple specific code paths being nominally O(N^3), dominating other per component costs when object and component counts get large enough.  There are 4 fundamental optimizations applied:

1.  Only run construction scripts once during an edit (outside the component inner loop), rather than per component in the loop.  This reduces one O(N^3) code path to O(N^2).
2.  Batch UObject archetype instance search function, which accepts an array of objects, allowing a single pass over the objects in the world.  Reduces a second O(N^3) code path to O(N^2).
3.  Generate Simple Construction Script (SCS) node map for the entirety of an edit.  Avoids O(N) node search in GetArchetype, called in a few places per component.  Reduces the worst case code paths from O(N^2) to O(N).
4.  Batch Add/Remove/ReleasePrimitive commands added to FScene, plus a context to gather and execute the batch commands.  This is the source of most of the level editor perf wins, but is also a large win for the blueprint code path.

Reference counting was added to the SCS node map, to allow nested code generating the node map, without inadvertently breaking the optimization if inner code removes the map (RerunConstructionScripts generates the node map, inside a higher level function that also uses it).  Note that the big perf wins in the blueprint editor apply even if there are no instances of the blueprint in the level in question, as most of the perf cost was within the blueprint editor itself, or iterating over the objects in the level, as opposed to the work done moving instance components that are actually found.  Some of the new batch calls may be useful for future optimizations elsewhere in the code base.

#jira UE-147154
#rnx
#rb ola.olsson danny.couture
#preflight 63d9d3689f6dbf248db62118

[CL 23976052 by jason hoerner in ue5-main branch]