UE-174107 - ExpandEnumAsExecs always uses first Exec if an argument name matches Enum value (resubmit)
ExpandEnumAsExecs was using an ambiguous pin lookup. These types of function nodes add additional exec pins that are named after the enumerators. If the target pin specified by ExpandEnumAsExecs happens to have the same name as an enumerator value, then we'll end up with the wrong pin if the basic FindPin function is used.

Instead, we prefer using FindPinByPredicate to specify the exact pin that we're interested in. This ensures that the expansion step creates a SwitchEnum node with a valid UEnum object.

#jira UE-174107
#preflight 63d9b16e65738ba9510f7d1e
#rb phillip.kavan

[CL 23983965 by dave jones2 in ue5-main branch]