Sequencer: Overhauled component inclusivity behavior / initialization This work is being done in order to re-introduce the concept of mutual component initializers in a more robust and manageable way. Mutally inclusive component rules define the presence (and now value) of a component based on the presence of another (or multiple). One example would be that DoubleResult[0] must be present alongside DoubleChannel[0].
Normally a mutual included component's presence is all that matters, but in some cases you want to initialize the value of that component to something based off the predicate component. This is a mutual component initializer.

Mutual component dependencies were previously stored as a digraph where edges denote inclusion. Additionally we supported 'complex' inclusions which were stored as an any/all component filter alongside the component mask to include. The logic for computing mutually inclusive components was previously implemented as a stabilizing loop algorithm that kept walking the graph and processing complex rules until it stopped adding components. This logic made it relatively hard to inject the notion of 'initializers' in a robust and performant way that correctly runs initializers if they were relevant to a particular operation (ie, we should not run an initializer if the component already existed).

The new logic stores simple and complex rules in separate maps, but has a generation step that generates a command buffer from the implied dependency graph. The resulting command buffer is ordered sequentially such that all rules that are predecated other mutually included components will be processed after their predicate has a chance to be included. This is decided based on the length of the predicate 'chain' in the graph. Cyclic dependencies are now an explicit error.

At runtime this command buffer is probably only generated once, and is relatively fast to execute (measured avg 600ns down from 12us on my Threadripper). It could be further optimized by removing commands that are known not to be relevant (ie, the leaf predicate component doesn't exist anywhere, so can never be active), but that is a separate task of diminishing returns.

Use of mutual component initializers are now mandatory when calling ComputeMutuallyInclusiveComponents. An option has been added to FEntityManager::MutateXYZ functions to specify whether you want to also add mutuals as part of the mutation.

Moved FMovieSceneEntitySystemDirectedGraph to UE::MovieScene::FDirectedGraph

#rb Max.Chen, Ludovic.Chabant
#preflight 63c8b8d12a6acaf16254ab65

[CL 24040366 by andrew rodham in ue5-main branch]