UnrealCloudDDC - Virtual assets permanent storage
We can now configure fallback namespaces, when a request for a blob in a namespace fails we can now fallback to another namespace. You will need to be authenticated for the second namespace as well. This allows us to configure virtual assets to query one namespace in the ddc pool (which is short lived) and if that fails it can fallback to the permanent VA storage, which is never GCed and only replicated on demand.

This also makes it possible to test the acl mapping even though auth is disabled.

#preflight 63e122d91b44ee7cb1da5da2

[CL 24047315 by Joakim Lindqvist in ue5-main branch]