Fix Vulkan crash when looking at hair in the Mesh LOD Coloration debug view mode.
Passing a pixel format when creating a view for a structured or byte address buffer confuses some RHIs and creates the wrong kind of view. We need to pass PF_Unknown in these cases. We're working on fixing this oddity in the RHI, but in the mean time this is an easy fix for the crash.

#jira UE-175996
#rnx
#rb Charles.deRousiers
#lockdown Juan.Canada
#preflight https://horde.devtools.epicgames.com/job/63e2217cc2257e56f403755c

[CL 24047434 by mihnea balta in ue5-main branch]