SourceControl: Fixed some changes not being immediately undone.
The batching of files into a single FCheckOut operation in #23509995 introduced a subtle bug if a user action resulted in two separate transactions within a single tick. An example of this is adding a component to an actor, which results in 'Add Component' and 'Clicking on Component (tree view)' transactions. Because the PreUnsavedAssetAutoCheckout notification got postponed to the Tick() because of the batching, only the 'Clicking on Component (tree view)' transaction ended up being undone. This fix restores the timing of the PreUnsavedAssetAutoCheckout to its original location and adds a PostUnsavedAssetAutoCheckoutCancel delegate which will be called for assets that do not end up being part of the CheckOutOperation (because they do not exist on disk, or are no longer dirty).

#preflight none

[CL 24066981 by wouter burgers in ue5-main branch]