Fixed crash when unloading animation assets in a game feature plugin that are referenced by editor thumbnails
Unloading a game feature plugin attempts to unload all of the assets therein. It attempts to GC unreferenced RF_Standalone assets in editor, but if the assets are still reachable it will leave them alone. Hoever the game feature plugin system then marks the 'leaking' assets with MarkAsGarbage. This causes weak references to those assets to return null, even though hard references in higher level code think they are valid. In this case, the animation decompression code was trying to use a skeleton that was marked as garbage.

In this case, assets were sill being referenced by the tick record tracking in the Sync member of the preview instance's proxy. The fix is to clear out the sync on setting animations and uninitializing the proxy so that the animations dont get picked up as reachable for GC.

[CL 24093144 by thomas sarkanen in ue5-main branch]