Horde: Fix for line numbers getting out of sync when processing bursty output.
It's possible that a log packet written to the server is truncated to stay below the maximum packet size, which results in some lines being retained in the log writer buffer. Code was previously assuming that subsequent packets would start at the next line index, causing the line count to be offset incorrectly. Since packets are flushed as soon as the threshold size is reached, it would commonly end up with one unused line in the buffer causing subsequent line counts to be offset by one.

#preflight none

[CL 24102435 by Ben Marsh in ue5-main branch]