Sequencer: fix incorrect setup and design for track areas
The pinned track area was shared with the normal track area, resulting in much painting and user interaction happening simultaneously in both areas. Now we properly create a pinned track area.

This change highlighted a problem which existed before the mvvm refactor, but became worse with that refactor: a lot of code needing access to the track area would simply get the "normal" track area from the sequencer editor, instead of getting the actual track area they are related to. This mean a lot of code was getting the "normal" track area when they were instead operating on the pinned one. Regressions have been fixed, but incorrect code from before the refactor has also been fixed. There might still be more to fix.

Before the mvvm refactor, the active hotspot was stored on the sequencer itself. Now, it's stored on the track area, which means there's no one place to figure out where the current hotspot is -- we would have to know which track area the mouse cursor is currently over. To avoid that, we bubble up the current hotspot from any of our track areas up to the sequencer editor view model. So now any code needing to simply know what is the current hotspot can look that up on the view model.

#jira UE-175785
#preflight 63e6dfd5211c20974f0b3d27
#rb max.chen

[CL 24179637 by ludovic chabant in ue5-main branch]