Path Tracing: Refactor path tracing code to minimize material/blend type related special cases in the integrator
Revisit payload encoding to handle blend mode logic entirely in the shader. This removes the need for a blendmode field and simplifies the logic in the raygen shader, reducing the amount of switch statements. We also gain back 3 extra bits that can be spent on output flags.

Refactor how random walk selection is performed to avoid shadingmodel specific code in the integrator. Also refactor the code to allow transparency to be combined with SSS.

Unify the handling of SingleLayerWater and solid glass since they were mostly sharing the same code. Both cases now use the DEFAULT_LIT shading model. This required extending the logic for camera medium tracking a bit. It now always runs (with a cvar to turn it off for debugging) instead of being tied to View.IsUnderwater() which could only work for water geometry. This makes it possible to achieve the under-water effect with regular refractive materials as well.

#rb Aleksander.Netzel
#jira none
#preflight 63efbaa090198dffbafaaf38

[CL 24286321 by chris kulla in ue5-main branch]