Fix conversions when using the "self" node.
When a pin type is PSC_Self, then its object class type is assumed to be that of the generated class. FindSpecializedConversionNode performed that check for the input pin, but *not* the output pin. This would prevent linking of the self node's pin to any input pin that would need a conversion (eg: linking the self node's pin to a soft reference pin).

If the output pin type is a self type, then the fix here simply changes the output pin type to use the generated class obtained via the input pin. Note that this assumes that the output pin's type also comes from a pin in the same Blueprint as the input.

#preflight 63f39837e832ea950a75ec89
#rb phillip.kavan

[CL 24344207 by dave jones2 in ue5-main branch]