Engine: Memory: MallocBinned3: Fixed a potential problem with double tracking of some physical allocations.
In the affected code we first allocate physical memory using one of the below:
1) GetCachedOSPageAllocator().Allocate     (if USE_CACHED_PAGE_ALLOCATOR_FOR_LARGE_ALLOCS == 1)
2) Block.Commit                                             (if USE_CACHED_PAGE_ALLOCATOR_FOR_LARGE_ALLOCS == 0)

and then increase the LLM Platform tracker. The problem is with case (1) because TCachedOSPageAllocator uses FPlatformMemory::BinnedAllocFromOS, which already handles LLM counter internally. I shoould rather say that BinnedAllosFromOS SHOULD update the LLM states, because it doesn't seem to be the case on HoloLens.So, we would account for the memory twice.

Case (2) works correctly because Block.Commit API allows arbitrary Commit/Decommit, which are not handled internally and it's the user code that has to do it.

Resolved by moving the LLM update code so that it only affected case (2). It was more of a potential rather than real issue because USE_CACHED_PAGE_ALLOCATOR_FOR_LARGE_ALLOCS (required for case (1) to be active) is set to 0 by default.

#preflight https://horde.devtools.epicgames.com/job/63f52a3a7064074bd8445b9f
#rb David.Harvey
#rnx

[CL 24412270 by Wojciech Krywult in ue5-main branch]