Fix incorrect task tag being applied by the parallel for body in BuildNaniteDrawCommands.
One of the tasks spawned by the parallel for will execute on the instigator thread, which is the render thread. Using FTaskTagScope will replace the existing tag with EParallelRenderingThread, which will trigger ensures in IsInRenderThread() & co, as the task tag won't match the thread identity. The correct approach is to use FOptionalTaskTagScope, which doesn't change the tag if one already exists.

I believe that FTaskTagScope doesn't have any valid uses and should be removed. All the other places which use it instead of FOptionalTaskTagScope seem like bugs waiting to happen.

#rnx
#jira UE-178460
#preflight none
#rb jeannoe.morissette, zach.bethel

[CL 24459367 by mihnea balta in ue5-main branch]