WinDualShock: Fixed DualSense controller being recognized as DualShock4 during the first update after launching the game.
The problem was caused by FWinDualShock::SendControllerEvents setting up a FInputDeviceScope before the controlller's type is initialized (which happens in SendControllerEvents).

An additional factor contributing to the issue is that while we do initialize controllers (and query their types) in FWinDualShock constructor by calling ConnectStateToUser, the type info is not available at this point. It's because controlllers' state is internally updated by libScePad library on a thread, so even if our controller is connected before we launch the game, it will simply return 'not connected' controller type, which we previously ignored.

Resolved by additionallly updating the controllers' types. The added benefit is that it fixed the problem of GetControllerTypeIdentifier and GetControllerTypeName returning stale type/name for controllers that have been disconnected.

#preflight https://horde.devtools.epicgames.com/job/63fe7107dd78dd50f62311fe
#rb ben.hoffman

#rn fix PS5 DualSense controllers used on PC are no longer incorrectly recognized as DualShock4 when they are initially connected. Previouslly, disconnected controllers would return DualShock4 as their type. Now, DualSense/DualShock4 types are only returned for controllers which state is 'connected'. Otherwise, None is returned for them.

[CL 24482164 by Wojciech Krywult in ue5-main branch]