It's now possible to register callbacks with a query for the Typed Elements Data Storage.
The Select query now accepts a callback which will be called every tick if there are any column combinations that match the query. Alternatively the callback can be triggered whenever a specific column is added or removed from a table. Information such as the required columns and dependencies are extracted automatically from the function signature.

#jira UE-171272
#rb logan.buchy
#preflight 64014106a20ddf1bf47345b1

[CL 24503688 by ronald koppers in ue5-main branch]