USDImporter: align engine-side prim IsA() and HasAPI() function signatures with USD API
As part of the work to support schema versioning in USD, a change went into v23.02
that refactors the IsA() and HasAPI() functions on UsdPrim. In particular with HasAPI(),
the instanceName parameter is only relevant when the schema being tested is a multiple
apply API schema, and the refactor makes it so that passing an empty instanceName
is now a coding error.

This change deprecates the single HasAPI() function that takes an optional instance name
parameter and replaces it with one- and two-FName parameter overloads that match the
one- and two-token counterparts on UsdPrim.

Using the engine's current v22.11 version of USD, we continue to get the TfType by its
schema identifier before passing it to the USD-side IsA() or HasAPI() functions that take
the schema by type. After the engine's USD version is upgraded to v23.02, we can instead
call the UsdPrim functions that take the schema identifier as a token directly and skip
getting the type ourselves.

More detail on the relevant change in USD is available in the commit here:

https://github.com/PixarAnimationStudios/USD/commit/5933c201ebc5eacb9198c615e30666e672e726c2

#jira UE-175005
#rb daniel.coelho, anousack.kitisa
#preflight 640b691470639dfc94e8b8d3

[CL 24594682 by matt johnson in ue5-main branch]