SharedMemoryMedia: Fixed sRGB handling.
The SMM player was not labeling the samples with the sRGB according to the format of the incoming shared texture. They were incorrectly always being flagged as sRGB encoded. This issue was exposed after fixing the sRGB handling in the nDisplay media input, which now triggers the correct conversion when the incoming textures are sRGB encoded.

#rb andrey.yamashev
#jira UE-179501
#preflight 640c0780c41a0a2a77eee976

[CL 24608991 by alejandro arango in ue5-main branch]