Moved stereo layer destruction logic to OnUnregister.
When the actor a stereo layer was attached to was destroyed the stereo layer wasn't destroyed. This is because the BeginDestroy callback is called by the GC at a future point in time which does not coincide with the actor destruction. On the other hand, components receive the OnUnregister callback when the actor is destroyed.

#jira UE-161758
#rb jeff.fisher
#preflight 640b6fa328026468d95f365b

[CL 24610573 by erica stella in ue5-main branch]