Sequencer: saving pre-animated state should be able to automatically start tracking the source if needed
The code was previously written mostly with object property tracks in mind, for which the tracking and caching of values is split between inputs (contributors to the value) and outputs (the property being read from and written to). For other tracks like the CVar track, however, there is no difference. We don't want people to have to explicitly start tracking sources, so this change adds an extra parameter to the pre-animated state caching APIs to auto-track sources if needed.

#jira UE-178389
#preflight 640f8c0b5819afacaf5c085f
#rb andrew.rodham

[CL 24623554 by ludovic chabant in ue5-main branch]