Add an idle timeout on perforce persistent connections
A connection to the perforce server can consume a large amount of memory on the server.  This change is to add a timeout to any persistent connections
in order to conserve this memory.
To do so, a connection tracks the platform time of the last successful communication.  The provider checks the persistent connection it has each frame and if
the last successful connect exceeded the timeout, then the connection is closed.  A new persistent connection will be established next time one is required.

The timeout is set as a hardcoded value for now.

#jira UE-173884
#rb Ronald.Koppers
#preflight 64122ee5af3fc352929f8e3b

[CL 24665411 by logan buchy in ue5-main branch]